tests/wpt/web-platform-tests/workers/interfaces/WorkerUtils/navigator/005.html

<!--
postMessage(navigator.userAgent);
/*
-->
<!doctype html>
<title>navigator.userAgent</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var worker = new Worker('#');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data, navigator.userAgent);
    this.done();
  });
});
</script>
<!--
*/
//-->
